pyproject.toml
[project]
name = "deriv-tech-interview"
version = "0.1.0"
description = "AI-powered incident analysis pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "langchain-anthropic>=1.4.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-openai>=1.2.1",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.1.10",
    "langsmith>=0.7.38",
    "pinecone>=6.0.0,<8.0.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.46.0",
]

[project.scripts]
run-pipeline = "main:main"
run-api = "api_main:main"
